feat: Organização MVC, Implementação de EstudanteServiço e Estrutura de Banco de Dados
