--- backend/src/Main.java
@@ -1,7 +1,0 @@
-import utils.Aplicacao;
-
-public class Main {
-    public static void main(String[] args) {
-        Aplicacao.iniciarAplicacao();
-    }
-}
