+++ backend/src/com/dbc/exceptions/EntradaUsuarioException.java
@@ -1,0 +1,7 @@
+package com.dbc.exceptions;
+
+public class EntradaUsuarioException extends Exception {
+    public EntradaUsuarioException(String mensagem) {
+        super(mensagem);
+    }
+}
