+++ backend/src/com/dbc/interfaces/IDocumentacaoProfissional.java
@@ -1,0 +1,6 @@
+package com.dbc.interfaces;
+
+public interface IDocumentacaoProfissional {
+    boolean validarDocumentos();
+    void adicionarCertificados(String certificados);
+}
