+++ backend/src/com/dbc/view/Main.java
@@ -1,0 +1,7 @@
+package com.dbc.view;
+
+public class Main {
+    public static void main(String[] args) {
+        Aplicacao.iniciarAplicacao();
+    }
+}
