--- backend/src/exceptions/EntradaUsuarioException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class EntradaUsuarioException extends Exception {
-    public EntradaUsuarioException(String mensagem) {
-        super(mensagem);
-    }
-}
