--- backend/src/interfaces/IDocumentacaoProfissional.java
@@ -1,6 +1,0 @@
-package interfaces;
-
-public interface IDocumentacaoProfissional {
-    boolean validarDocumentos();
-    void adicionarCertificados(String certificados);
-}
